% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_batch_jobs.R
\name{use_batch_jobs}
\alias{use_batch_jobs}
\title{Create a jobs directory within the specified path}
\usage{
use_batch_jobs(path, template = c("rmarkdown", "quarto"))
}
\arguments{
\item{path}{Directory for which jobs directory should be created}

\item{template}{Whether to use R Markdown, quarto, or both types of templates}
}
\value{
Path of the job directory created
}
\description{
This function creates a directory named 'jobs' and copies template files into
the newly created directory. Either R Markdown, Quarto, or both types of
files can be used as templates.
}
